kb.c: I made a stupid mistake, thanks Evandro, classify.c , just revert, too tired to fix and hate VC too much.
git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@4214 c60273ff-4a12-0410-8d58-82ceea6d4170

+++ sphinx3/src/libs3decoder/libep/classify.c
@@ -279,7 +279,7 @@ void calclikeli (mgau_model_t *g, /*Input: multiptle mixture models */
 
 	  b += (f - m) * (f - m) * v; 
 	}	
-	pmix = a * exp ((float)(-0.5) * b);	
+	pmix = a * exp (-0.5 * b);	
 	likeli[i] += g->mgau[i].mixw_f[j] * pmix; 
       }	
   }
